requirements.txt
absl-py==1.4.0
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.1.1
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beem==0.24.26
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.0.3
click-shell==2.1
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
deluge==2.0.3
diff-match-patch==20230430
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
ecdsa==0.18.0
evdev==1.4.0
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.15.4
fire==0.5.0
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2024.6.1
future==0.18.2
gast==0.4.0
GeoIP==1.3.2
google-api-core==2.19.2
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
gpg==1.16.0
gradio_client==1.3.0
grpcio==1.66.1
grpcio-status==1.66.1
gyp==0.1
h11==0.14.0
h5py==3.8.0
hidpidaemon==18.4.6
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.24.6
hyperlink==21.0.0
idna==2.8
importlib-metadata==4.6.4
incremental==21.3.0
jax==0.4.8
jeepney==0.7.1
Jinja2==3.0.3
keras==2.12.0
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
libtorrent===2.0.5-build-libtorrent-rasterbar-qrM5vM-libtorrent-rasterbar-2.0.5-bindings-python
livereload==2.6.3
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.2
mkdocs==1.1.2
ml-dtypes==0.1.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
netaddr==0.8.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
paramiko==2.9.3
pdfkit==1.0.0
pexpect==4.8.0
pillow==10.2.0
pipx==1.2.0
pop-transition==1.1.2
prettytable==3.8.0
proto-plus==1.24.0
protobuf==5.28.0
protontricks==1.11.0
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.53
pycairo==1.20.1
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.18.0
pycups==2.0.1
pydbus==0.6.0
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.6
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
Quamash==0.6.1
regex==2017.4.5
rencode==1.0.6
repolib==2.2.1
repoman==1.4.0
requests==2.21.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.4.4
scipy==1.10.1
scrypt==0.8.20
SecretStorage==3.3.1
service-identity==18.1.0
sessioninstaller==0.0.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
sympy==1.12
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tokenizers==0.19.1
toposort==1.10
torbrowser-launcher==0.3.3
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
tornado==6.1
tqdm==4.66.5
transformers==4.44.2
Twisted==22.1.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.24.3
userpath==1.8.0
vdf==3.4
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.2
websockets==12.0
Werkzeug==2.3.3
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.8.2
you-get==0.4.1650
zipp==1.0.0
zope.interface==5.4.0
